+++ src/archive_item.hpp
@@ -1,0 +1,30 @@
+#pragma once
+
+#include <filesystem>
+#include <string>
+
+namespace hisui {
+
+class ArchiveItem {
+ public:
+  ArchiveItem(const std::filesystem::path&,
+              const std::string&,
+              const double,
+              const double);
+
+  std::filesystem::path getPath() const;
+  std::string getConnectionID() const;
+  double getStartTimeOffset() const;
+  double getStopTimeOffset() const;
+  void adjustTimeOffsets(double);
+
+  ArchiveItem& operator=(const ArchiveItem& other);
+
+ private:
+  std::filesystem::path m_path;
+  std::string m_connection_id;
+  double m_start_time_offset;
+  double m_stop_time_offset;
+};
+
+}  // namespace hisui