+++ test/version/version_test.cpp
@@ -5,7 +5,7 @@
 BOOST_AUTO_TEST_SUITE(version)
 
 BOOST_AUTO_TEST_CASE(get_hisui_version) {
-  BOOST_REQUIRE_EQUAL("2021.3", hisui::version::get_hisui_version());
+  BOOST_REQUIRE_EQUAL("2022.1.0", hisui::version::get_hisui_version());
 }
 
 BOOST_AUTO_TEST_SUITE_END()