+++ CHANGES.md
@@ -12,13 +12,13 @@
 ## develop
 
 - [UPDATE] 依存ライブラリの更新
-    - `boost` を `1.77.0` に
+    - `boost` を `1.78.0` に
     - `CLI11` を `2.1.2` に
     - `fmt` を `8.0.1` に
     - `spdlog` を `1.9.2` に
     - `rapidcsv` を `8.53` に
     - `libvpx` を `v.1.11.0` に
-    - `cpp-mp4` を `7978dcff455a805c8ef941a45ac5aba1941fe57d` に
+    - `cpp-mp4` を `252fd4f0104d072a1a44b4e3ba78126cf64d0103` に
     - `stb` を `af1a5bc352164740c1cc1354942b1c6b72eacb8a` に
     - @haruyama